# gm-center/gmc_api_gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Gedge GM-Center Swagger API.",
        "title": "Gedge GM-Center Swagger API",
        "contact": {
            "name": "GM-Center",
            "url": "https://gedge-platform.github.io/gm-center/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "101.79.1.138:8012",
    "basePath": "/gmcapi/v2",
    "paths": {
...
        "/statefulsets": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get Statefulset List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Show List Statefulset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.WORKLOAD"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create Statefulset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Create Statefulset",
                "parameters": [
                    {
                        "description": "Statefulset Info Body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/statefulsets/{name}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get Statefulset Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Show detail Statefulset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Statefulset",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.STATEFULSET_DETAIL"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete Statefulset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Delete Statefulset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Statefulset",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.EVENT": {
            "type": "object",
            "properties": {
                "cluster": {
                    "type": "string"
                },
                "eventTime": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.STATEFULSET_DETAIL": {
            "type": "object",
            "properties": {
                "NodeSelector": {},
                "annotations": {},
                "cluster": {
                    "type": "string"
                },
                "containers": {},
                "createAt": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EVENT"
                    }
                },
                "label": {},
                "name": {
                    "type": "string"
                },
                "ownerReferences": {},
                "project": {
                    "type": "string"
                },
                "ready": {
                    "type": "string"
                },
                "status": {},
                "user": {
                    "type": "string"
                },
                "workspace": {
                    "type": "string"
                }
            }
        },
...
        "model.WORKLOAD": {
            "type": "object",
            "properties": {
                "NodeSelector": {},
                "cluster": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "ready": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                },
                "workspace": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type \\\"Bearer \\\" and then your API Token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}